+++ starsky/starskytest/starsky.foundation.storage/Storage/StorageHostFullPathFilesystemTest.cs
@@ -1,4 +1,7 @@
+using System;
 using System.IO;
+using System.Linq;
+using System.Reflection;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.foundation.storage.Models;
 using starsky.foundation.storage.Storage;
@@ -9,6 +12,19 @@ namespace starskytest.starsky.foundation.storage.Storage
 	[TestClass]
 	public class StorageHostFullPathFilesystemTest
 	{
+		[TestMethod]
+		public void Files_GetFilesRecrusiveTest()
+		{            
+			var path = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location) + Path.DirectorySeparatorChar;
+
+			var content = new StorageHostFullPathFilesystem().GetAllFilesInDirectoryRecursive(path);
+
+			Console.WriteLine("count => "+ content.Count());
+
+			// Gives a list of the content in the temp folder.
+			Assert.AreEqual(true, content.Any());            
+		}
+		
 		[TestMethod]
 		public void InfoNotFound()
 		{