+++ history.md
@@ -32,14 +32,17 @@ node starsky-tools/build-tools/app-version-update.js
 # Folder and file movable, Epic (Work In Progress)
 - [ ]   (x) Move multiple files __not implemented__
 
-# version 0.4.0 _(Unreleased)_ -  2020-11-??
+# version 0.4.1 _(Unreleased)_ - 2020-11-??
+- nothing yet
+
+# version 0.4.0 - 2020-11-14
 _Please check the breaking changes of 0.4.0-beta.0 and 0.4.0-beta.1_
 - [x]   (Changed) _App_ Add styling to settings UI in App
 - [x]   (Fixed) _Back-end_  Add extra catch to prevent sync issues when exif reading fails
 - [x]   (Deprecated) _Back-end_ Json Sidecar format is very likely to change in future releases and be incompatible
 - [x]   (Added) _App_ Add extra delay to check for updates to avoid issues when local
 - [x]   (Added) _App_ Add fix for selecting wrong domains to avoid an exception
-- [x]   (Fixed) _Back-end_ When switching very fast after update, info isn't updated until process is done
+- [x]   (Fixed) _Back-end_ When switching very fast after update, info isn't updated until process is done (this is fixed)
 - [x]   (Security) _Back-end_  Upgrade .NET Core (TargetFramework) to 3.1.9 (using SDK 3.1.403)
 - [x]   (Fixed) _Front-end_ Clean Front-end cache when moving file/renaming file
 - [x]   (Fixed) _Front-end_ Change text when selecting an non existing filter combination
@@ -47,8 +50,9 @@ _Please check the breaking changes of 0.4.0-beta.0 and 0.4.0-beta.1_
 - [x]   (Fixed) _Back-end_ Allow upload to folder with files that are uppercase
 - [x]   (Fixed) _Back-end_ Database-item is now correct updated when you move an item to the root folder (/)
 - [x]   (Security) _App_ Update Electron to 10.1.5 (Node 12.16.x and Chromium 85.0.x)
-- [x]   (Added) _Back-end_ In the rename API When enable Collections, this files are also moved
+- [x]   (Added) _Back-end_ In the rename/move API When enable Collections, this files are also moved (file to folder)
 - [x]   (Added) _Back-end_ Xmp sidecar files are moved with gif/bmp/Raw/mp4 file types
+- [x]   (Added) _Back-end_ In the rename API When enable Collections, this files are also moved (file to deleted)
 
 # version 0.4.0-beta.2 - 2020-11-04
 - [x]   (Changed) _Front-end_ Enable sockets client side option by default